requirements.txt
Django==5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
google-cloud-vision==3.5.0
google-generativeai==0.3.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
boto3==1.34.19
django-storages==1.14.2
